pyproject.toml
[tool.poetry]
name = "audioverba"
version = "0.1.0"
description = "Desktop audio transcription tool powered by Reverb ASR."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"  # Updated from >=3.9
pyside6 = "^6.0"  # Changed from pyside2
rev-reverb = "^0.1.0"
soundfile = "^0.13.1"
pyannote-audio = "^3.3.2"
huggingface-hub = "^0.30.2"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
